data-agent/pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "Course-style NL2SQL data agent backend."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "asyncmy",
  "cryptography",
  "elasticsearch[async]>=8,<9",
  "fastapi[standard]",
  "greenlet",
  "httpx",
  "jieba",
  "langchain",
  "langchain-huggingface",
  "langchain-openai",
  "langgraph",
  "loguru",
  "omegaconf",
  "pyyaml",
  "qdrant-client",
  "sqlalchemy",
]

[dependency-groups]
dev = [
  "httpx",
  "pytest",
  "pytest-asyncio",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
